refactor: Move components into specific folders
